% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortIncidenceQueryBuilder.R
\name{buildOptions}
\alias{buildOptions}
\title{Builds the BuilderOptions jObject with the specified paramaters}
\usage{
buildOptions(
  cohortTable,
  outcomeCohortTable = cohortTable,
  subgroupCohortTable = cohortTable,
  sourceName,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  vocabularySchema = cdmDatabaseSchema,
  useTempTables = FALSE,
  refId
)
}
\arguments{
\item{cohortTable}{The name of table with cohorts}

\item{outcomeCohortTable}{The name of table with outcome cohorts, defaults to cohortTable param.}

\item{subgroupCohortTable}{The name of table with subgroup cohorts, defaults to cohortTable param.}

\item{sourceName}{A value to inject to the results table for the source name.}

\item{cdmDatabaseSchema}{the name of schema containing data in CDM format}

\item{resultsDatabaseSchema}{the name of schema where results would be placed}

\item{vocabularySchema}{the name of schema with vocabulary tables, defaults to cdmDatabaseSchema param}

\item{useTempTables}{use temp tables instead of a results schema.}

\item{refId}{A number tagged to the results for retrieval purposes.}
}
\value{
a BuilderOptions object used in buildQuery.
}
\description{
Builds the BuilderOptions jObject with the specified paramaters
}
