% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\docType{package}
\name{CohortIncidence-package}
\alias{CohortIncidence}
\alias{CohortIncidence-package}
\title{CohortIncidence: Cohort Incidence Analysis for the OMOP Common Data Model}
\description{
Provides functionality for performing cohort incidence analyses against data stored in the OMOP Common Data Model (CDM). The package generates database-specific SQL, executes analyses across supported database platforms, and returns standardized incidence estimates using configurable time-at-risk, stratification, and cohort definitions. It is intended to support reproducible observational research within the OHDSI analytics framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortIncidence/}
  \item Report bugs at \url{https://github.com/OHDSI/CohortIncidence/issues}
}

}
\author{
\strong{Maintainer}: Christopher Knoll \email{cknoll@ohdsi.org}

Authors:
\itemize{
  \item Christopher Knoll \email{cknoll@ohdsi.org}
}

}
\keyword{internal}
