% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createCohortSubgroup}
\alias{createCohortSubgroup}
\title{Creates R6 object for CohortSubgroup}
\usage{
createCohortSubgroup(id, name, description, cohortRef)
}
\arguments{
\item{id}{the unique identifier for this subgroup}

\item{name}{The subgroup name}

\item{description}{The subgroup description}

\item{cohortRef}{A cohort reference, as an R6 Class CohortReference}
}
\value{
A \code{CohortSubgroup} R6 object containing subgroup metadata and the \code{CohortReference} used to determine subgroup membership.
}
\description{
Creates R6 object for CohortSubgroup
}
