% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createDateRange}
\alias{createDateRange}
\title{Creates R6 object for DateRange}
\usage{
createDateRange(startDate, endDate)
}
\arguments{
\item{startDate}{a character vector representing a date in YYYY-MM-DD format}

\item{endDate}{a character vector representing a date in YYYY-MM-DD format}
}
\value{
A \code{DateRange} R6 object containing the start and end date strings in \code{YYYY-MM-DD} format.
}
\description{
Creates R6 object for DateRange
}
