% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createTimeAtRiskDef}
\alias{createTimeAtRiskDef}
\title{Creates R6 object for TimeAtRisk}
\usage{
createTimeAtRiskDef(
  id,
  startWith = "start",
  startOffset = 0,
  endWith = "end",
  endOffset = 0
)
}
\arguments{
\item{id}{the unique identifier for this outcome definition}

\item{startWith}{Specifies the field (start or end) to offset from for the TAR start.  Allowed values: 'start', 'end'}

\item{startOffset}{The number of days to offset for the TAR start, defaults to 0.}

\item{endWith}{Specifies the field (start or end) to offset from for the TAR end.  Allowed values: 'start', 'end'}

\item{endOffset}{The number of days to offset for the TAR start, defaults to 0.}
}
\value{
A \code{TimeAtRisk} R6 object describing the TAR start and end anchors plus their offsets. The serialized form contains a \code{start} and \code{end} field, each with a \code{dateField} and \code{offset}.
}
\description{
Creates R6 object for TimeAtRisk
}
