% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortIncidence.R
\name{executeAnalysis}
\alias{executeAnalysis}
\title{Executes IR analysis given a design, options, and connection settings.}
\usage{
executeAnalysis(
  connectionDetails = NULL,
  connection = NULL,
  incidenceDesign,
  buildOptions,
  sourceName = "default"
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package. Either the \code{connection} or
\code{connectionDetails} argument should be specified.}

\item{connection}{A connection to the server containing the schema as created using the
\code{connect} function in the \code{DatabaseConnector} package.
Either the \code{connection} or \code{connectionDetails} argument
should be specified.}

\item{incidenceDesign}{A string object containing valid JSON that represents cohort incidence design}

\item{buildOptions}{the parameters to use in building the analysis queries, created by buildOptions()}

\item{sourceName}{the source name to attach to the results}
}
\value{
A named list of \code{data.frame} objects containing the IR results. The list includes \code{incidence_summary}, \code{target_def}, \code{outcome_def}, \code{tar_def}, \code{age_group_def}, and \code{subgroup_def}, each representing one result table produced during execution.
}
\description{
Executes IR analysis given a design, options, and connection settings.
}
\details{
This method performs the entire cohort incidence analysis in one step without the need for creating any permenant tables.

The process for calculating the cohort incidence is as follows:
\itemize{
\item Create the time at risk episodes from the target cohorts, using the specified startWith, endWith and offsets
\item Calculate the excluded time by finding the overlap of TAR with the following:
\item 
\enumerate{
\item Outcome cohort episodes (adding the clean window to each outcome episode end date)
}
\item 
\enumerate{
\item Exclusion cohort that was specified in the outcome definition excludeCohortId.
}
\item Count the distinct persons, distinct persons with cases, total time at risk and total cases, and calcuate the proportion and rates.
}

The resut contains the following dataframes:

incidence_summary:\tabular{ll}{
   Name \tab Description \cr
   REF_ID \tab The reference id specified in buildOptions() to track results to the analysis execution. \cr
   SOURCE_NAME \tab The name of the source for these results \cr
   TARGET_COHORT_DEFIITION_ID \tab The cohort ID of the target population \cr
   TAR_ID \tab The TAR identifier \cr
   SUBGROUP_ID \tab The subgroup identifier \cr
   OUTCOME_ID \tab The outcome identifier \cr
   AGE_GROUP_ID \tab The age ID for this strata representing the age band specified in the strata settings \cr
   GENDER_ID \tab The gender concept ID for this gender strata \cr
   GENDER_NAME \tab The name of the gender \cr
   START_YEAR \tab The year strata, defined by using the year the TAR started \cr
   PERSONS_AT_RISK_PE \tab Distinct persons at risk before removing excluded time from TAR \cr
   PERSONS_AT_RISK \tab Distinct persons at risk after removing excluded time from TAR.  A person must have at least 1 day TAR to be included. \cr
   PERSON_DAYS_PE \tab Total TAR (in days) before excluded time was removed from TAR. \cr
   PERSON_DAYS \tab Total TAR (in days) after excluded time was removed from TAR. \cr
   PERSON_OUTCOMES_PE \tab Distinct persons with outcome before removing excluded time from TAR \cr
   PERSON_OUTCOMES \tab Distinct persons with outcome after removing excluded time from TAR.  A person must have at least 1 day TAR to be included. \cr
   OUTCOMES_PE \tab Number of cases before excluding TAR. \cr
   OUTCOMES \tab Number of cases after excluding TAR. \cr
   INCIDENCE_PROPORTION_P100P \tab The Incidence Proportion (per 100 people), calculated by person_outcomes / persons_at_risk * 100 \cr
   INCIDENCE_RATE_P100PY \tab The Incidence Rate (per 100 person years), calculated by outcomes / person_days / 365.25 * 100 \cr
   target_def: \tab  \cr
   Name \tab Description \cr
   ----- \tab -------- \cr
   REF_ID \tab The reference id specified in buildOptions() to track results to the analysis execution. \cr
   TARGET_COHORT_DEFIITION_ID \tab The cohort ID of the target population \cr
   TARGET_NAME \tab The name of the target cohort \cr
}


outcome_def:\tabular{ll}{
   Name \tab Description \cr
   REF_ID \tab The reference id specified in buildOptions() to track results to the analysis execution. \cr
   OUTCOME_ID \tab The outcome identifier \cr
   OUTCOME_COHORT_DEFINITION_ID \tab The cohort ID of the outcome population \cr
   OUTCOME_NAME \tab The outcome name \cr
   CLEAN_WINDOW \tab The clean window for this outcome definition \cr
   EXCLUDED_COHORT_DEFINITION_ID \tab The cohort used to exclude TAR \cr
}


tar_def:\tabular{ll}{
   Name \tab Description \cr
   REF_ID \tab The reference id specified in buildOptions() to track results to the analysis execution. \cr
   TAR_ID \tab The TAR identifier \cr
   TAR_START_WITH \tab Indicates if the TAR starts with the 'start' or 'end' of the target cohort episode \cr
   TAR_START_OFFSET \tab The days added to the date field specified in TAR_START_WITH \cr
   TAR_END_WITH \tab Indicates if the TAR ends with the 'start' or 'end' of the target cohort episode \cr
   TAR_END_OFFSET \tab The days added to the date field specified in TAR_END_WITH \cr
}


age_group_def:\tabular{ll}{
   Name \tab Description \cr
   REF_ID \tab The reference id specified in buildOptions() to track results to the analysis execution. \cr
   AGE_GROUP_ID \tab The age ID for this strata representing the age band specified in the strata settings \cr
   AGE_GROUP_NAME \tab The name for this age group \cr
   MIN_AGE \tab The minimum age for this group \cr
   MAX_AGE \tab the max age for this group \cr
}


subgroup_def:\tabular{ll}{
   Name \tab Description \cr
   REF_ID \tab The reference id specified in buildOptions() to track results to the analysis execution. \cr
   SUBGROUP_NAME \tab The name of the subgroup \cr
}
}
