% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getCleanupSql}
\alias{getCleanupSql}
\title{Gets the sql that drops the results tables}
\usage{
getCleanupSql(useTempTables = FALSE)
}
\arguments{
\item{useTempTables}{if TRUE, then temp table notation will be used.}
}
\value{
SQL code in MS Sql Server dialect. If analysis will run on another DBMS, translate the SQL with \code{SqlRender::translateSql()} before execution.
}
\description{
Gets the sql that drops the results tables
}
