% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortIncidenceQueryBuilder.R
\name{buildOptions}
\alias{buildOptions}
\title{Builds the BuilderOptions jObject with the specified paramaters}
\usage{
buildOptions(
  cohortTable,
  outcomeCohortTable = cohortTable,
  subgroupCohortTable = cohortTable,
  databaseName,
  cdmSchema,
  resultsSchema,
  vocabularySchema = cdmSchema,
  tempSchema = resultsSchema,
  refId
)
}
\arguments{
\item{cohortTable}{The name of table with cohorts}

\item{cdmSchema}{the name of schema containing data in CDM format}

\item{resultsSchema}{the name of schema where results would be placed}

\item{vocabularySchema}{the name of schema with vocabularies}

\item{tempSchema}{the name of database temp schema}

\item{sessionId}{session identifier using to build temporary tables}

\item{analysisId}{analysis identifier}
}
\value{
a BuilderOptions object used in buildQuery.
}
\description{
Builds the BuilderOptions jObject with the specified paramaters
}
