% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortIncidenceQueryBuilder.R
\name{buildQuery}
\alias{buildQuery}
\title{Builds SQL code to run analyses according given Cohort Characterization design}
\usage{
buildQuery(incidenceDesign, buildOptions)
}
\arguments{
\item{incidenceDesign}{A string object containing valid JSON that represents cohort incidence design}

\item{buildOptions}{the paramaters to use in building the analysis queries, created by buildOptions()}
}
\value{
SQL code in MS Sql Server dialect, if it's required to run analysis on another DBMS
        you have to use \code{\link[SqlRender]{translateSql}} function in the SqlRender package.
}
\description{
Builds SQL code to run analyses according given Cohort Characterization design
}
