% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\docType{package}
\name{CohortIncidence-package}
\alias{CohortIncidence}
\alias{CohortIncidence-package}
\title{CohortIncidence: CohortIncidence Package}
\description{
An R package that produces SQL queries for executing Cohort Incidence analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/CohortIncidence}
  \item Report bugs at \url{https://github.com/OHDSI/CohortIncidence/issues}
}

}
\keyword{internal}
