% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{CohortSubgroup}
\alias{CohortSubgroup}
\title{R6 Class Representing a Cohort Subgroup definition}
\description{
The CohortSubgroup class, encapsulating the id, name, description, and CohortRef.
}
\details{
This class is used to specify a cohort subgroup to be used in the analysis. A TAR will be considered
part of the subgroup if the TAR starts between the subgroup's cohort start and cohort end.

Note, when serializing with a library such as jsonlite, first call toList() on the R6 class
before calling jsonlite::toJSON().
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{an integer uniquely identifying this subgroup}

\item{\code{name}}{The name to use for this cohort reference}

\item{\code{description}}{The description for this subgroup}

\item{\code{cohort}}{the cohort used to represent this subgroup.  Must be class CohortReference}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CohortSubgroup$new()}}
\item \href{#method-toList}{\code{CohortSubgroup$toList()}}
\item \href{#method-asJSON}{\code{CohortSubgroup$asJSON()}}
\item \href{#method-clone}{\code{CohortSubgroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
creates a new instance, using the provided data param if provided.
The JSON takes the form: {"id":1,"name":"some name","description":"some description","cohort":{"id":99, "name":"cohort"}}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubgroup$new(data = list(CohortSubgroup = list()))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{the data (as a json string or list) to initialize with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toList"></a>}}
\if{latex}{\out{\hypertarget{method-toList}{}}}
\subsection{Method \code{toList()}}{
returns the R6 class elements as a list for use in jsonlite::toJSON()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubgroup$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asJSON"></a>}}
\if{latex}{\out{\hypertarget{method-asJSON}{}}}
\subsection{Method \code{asJSON()}}{
returns the JSON string for this R6 class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubgroup$asJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubgroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
