% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createCohortRef}
\alias{createCohortRef}
\title{Creates R6 object for CohortReference}
\usage{
createCohortRef(id, name, description)
}
\arguments{
\item{id}{the cohort id that is being referenced}

\item{name}{the name to use for this reference}

\item{description}{an optional description to use for this reference.}
}
\value{
a R6 class: CohortReference
}
\description{
Creates R6 object for CohortReference
}
