% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createCohortRef}
\alias{createCohortRef}
\title{Helper function for creating Cohort References, which are used in different parts of a design to reference a cohort definition
and, optionally, to provide a name.}
\usage{
createCohortRef(cohortId, name, description)
}
\arguments{
\item{cohortId}{the cohort id that is being referenced}

\item{name}{the name to use for this reference}

\item{description}{an optional description to use for this reference.}
}
\value{
an R list containing name-value pairs that will serialize into a org.ohdsi.analysis.CohortRef JSON format.
}
\description{
Helper function for creating Cohort References, which are used in different parts of a design to reference a cohort definition
and, optionally, to provide a name.
}
