% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createCohortSubgroup}
\alias{createCohortSubgroup}
\title{Helper function for creating TAR Definitions}
\usage{
createCohortSubgroup(id, name, description, cohortRef)
}
\arguments{
\item{id}{the unique identifier for this subgroup}

\item{name}{The subgroup name}

\item{description}{The subgroup description}

\item{cohortRef}{A cohort reference created by calling createCohortRef()}
}
\value{
an R list containing name-value pairs that will serialize into a org.ohdsi.analysis.cohortincidence.design.CohortSubgroup JSON format.
}
\description{
Helper function for creating TAR Definitions
}
