% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createIncidenceAnalysis}
\alias{createIncidenceAnalysis}
\title{Creates R6 object for IncidenceAnalysis}
\usage{
createIncidenceAnalysis(targets, outcomes, tars)
}
\arguments{
\item{targets}{A list or vector of target IDs from target definitions.}

\item{outcomes}{A list or vector of outcome IDs from outcome definitions.}

\item{tars}{A list or vector of TAR IDs from time-at-risk definitions.}
}
\value{
a R6 class: IncidenceAnalysis
}
\description{
Creates R6 object for IncidenceAnalysis
}
