% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createIncidenceDesign}
\alias{createIncidenceDesign}
\title{Creates R object for Incidence Design}
\usage{
createIncidenceDesign(
  cohortDefs = list(),
  targetDefs = list(),
  outcomeDefs = list(),
  tars = list(),
  analysisList = list(),
  conceptSets = list(),
  subgroups = list()
)
}
\value{
SQL code in MS Sql Server dialect, if it's required to run analysis on another DBMS
        you have to use \code{\link[SqlRender]{translateSql}} function in the SqlRender package.
}
\description{
Creates R object for Incidence Design
}
