% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createIncidenceDesign}
\alias{createIncidenceDesign}
\title{Creates R object for Incidence Design}
\usage{
createIncidenceDesign(
  cohortDefs = list(),
  targetDefs = list(),
  outcomeDefs = list(),
  tars = list(),
  analysisList = list(),
  conceptSets = list(),
  subgroups = list()
)
}
\arguments{
\item{cohortDefs}{The set of cohort definitions.  Optional.}

\item{targetDefs}{A list of target definitions, each created with createCohortRef.}

\item{outcomeDefs}{A list of outcome definitions, each created with createOutcomeDef.}

\item{tars}{A list of TAR definitions, each created with createTimeAtRiskDef.}

\item{analysisList}{A list of analysis definitions, each created with createIncidenceAnalysis.}

\item{conceptSets}{A list of concept sets, currently unused.}

\item{subgroups}{A list of cohort subgroups, creatd with createCohortSubgroup.}
}
\value{
SQL code in MS Sql Server dialect, if it's required to run analysis on another DBMS
        you have to use \code{\link[SqlRender]{translateSql}} function in the SqlRender package.
}
\description{
Creates R object for Incidence Design
}
