% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createIncidenceDesign}
\alias{createIncidenceDesign}
\title{Creates R6 object for IncidenceDesign}
\usage{
createIncidenceDesign(
  cohortDefs,
  targetDefs,
  outcomeDefs,
  tars,
  analysisList,
  conceptSets,
  subgroups,
  strataSettings,
  studyWindow
)
}
\arguments{
\item{cohortDefs}{The set of cohort definitions.  Optional.}

\item{targetDefs}{A list of target definitions, each element must be class CohortReference.}

\item{outcomeDefs}{A list of outcome definitions, each element must be class Outcome}

\item{tars}{A list of TAR definitions, each element must be class TimeAtRisk}

\item{analysisList}{A list of analysis definitions, each element must be class IncidenceAnalysis}

\item{conceptSets}{A list of concept sets, currently unused.}

\item{subgroups}{A list of cohort subgroups, each element must be class Subgroup.}

\item{strataSettings}{The strata settings used in the anlaysis, must be class StrataSettings.}

\item{studyWindow}{Limits time at risk to the specified study window. Must be class DateRange.}
}
\value{
a R6 class: IncidenceDesign.
}
\description{
Creates R6 object for IncidenceDesign
}
