% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createOutcomeDef}
\alias{createOutcomeDef}
\title{Helper function for creating Outcome Definitions}
\usage{
createOutcomeDef(id, name, cohortId = 0, cleanWindow = 0, excludeCohortId)
}
\arguments{
\item{id}{the unique identifier for this outcome definition}

\item{name}{an optional name for this outcome definition}

\item{cohortId}{the cohort id reference for this outcome}

\item{cleanWindow}{the number of days for the clean window of this outcome definition}

\item{excludeCohortId}{a cohort ID for the cohort to use to exclude time at risk}
}
\value{
an R list containing name-value pairs that will serialize into a org.ohdsi.analysis.cohortincidence.design.Outcome JSON format.
}
\description{
Helper function for creating Outcome Definitions
}
