% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createStrataSettings}
\alias{createStrataSettings}
\title{Helper function for creating strata settings}
\usage{
createStrataSettings(byAge = F, byGender = F, byYear = F, ageBreaks = NULL)
}
\arguments{
\item{byAge}{a boolean indicating to stratify by age}

\item{byGender}{a boolean indicating to stratify by gender}

\item{byYear}{a boolean indicating to stratify by year}

\item{ageBreaks}{a vector of integers indicating the age group bounds.}
}
\value{
an R list containing name-value pairs that will serialize into a org.ohdsi.analysis.cohortincidence.design.StratifySettings JSON format.
}
\description{
Helper function for creating strata settings
}
