% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builders.R
\name{createStrataSettings}
\alias{createStrataSettings}
\title{Creates R6 object for StrataSettings}
\usage{
createStrataSettings(
  byAge = F,
  byGender = F,
  byYear = F,
  ageBreaks,
  ageBreakList
)
}
\arguments{
\item{byAge}{a boolean indicating to stratify by age, defaults to F}

\item{byGender}{a boolean indicating to stratify by gender, defaults to F}

\item{byYear}{a boolean indicating to stratify by year, defaults to F}

\item{ageBreaks}{a vector of integers indicating the age group bounds.}
}
\value{
an R list containing name-value pairs that will serialize into a org.ohdsi.analysis.cohortincidence.design.StratifySettings JSON format.
}
\description{
Creates R6 object for StrataSettings
}
