% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{DateRange}
\alias{DateRange}
\title{R6 Class Representing a DataRange}
\value{
A \code{DateRange} R6 object storing \code{startDate} and \code{endDate} as \code{YYYY-MM-DD} strings.
}
\description{
The DateRange class, encapsulating the startDate and endDate of a date range.
}
\details{
This class is used to specify a DateRange, with start and end dates specified as
strings formatted as YYYY-MM-DD.
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{startDate}}{a character with format YYYY-MM-DD to be used as the date range's start date.}

    \item{\code{endDate}}{a character with format YYYY-MM-DD to be used as the date range's end date.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DateRange-initialize}{\code{DateRange$new()}}
    \item \href{#method-DateRange-toList}{\code{DateRange$toList()}}
    \item \href{#method-DateRange-asJSON}{\code{DateRange$asJSON()}}
    \item \href{#method-DateRange-clone}{\code{DateRange$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DateRange-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DateRange-initialize}{}}}
\subsection{\code{DateRange$new()}}{
  creates a new instance, using the provided data param if provided.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DateRange$new(data = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{the data (as a json string or list) to initialize with}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{DateRange} object initialized from a list or JSON string.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DateRange-toList"></a>}}
\if{latex}{\out{\hypertarget{method-DateRange-toList}{}}}
\subsection{\code{DateRange$toList()}}{
  returns the R6 class elements as a list for use in jsonlite::toJSON()
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DateRange$toList()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A named list with \code{startDate} and \code{endDate} fields.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DateRange-asJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DateRange-asJSON}{}}}
\subsection{\code{DateRange$asJSON()}}{
  returns the JSON string for this R6 class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DateRange$asJSON()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A JSON string representation of the \code{DateRange} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DateRange-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DateRange-clone}{}}}
\subsection{\code{DateRange$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DateRange$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
