% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{DateRange}
\alias{DateRange}
\title{R6 Class Representing a DataRange}
\description{
The DateRange class, encapsulating the startDate and endDate of a date range.
}
\details{
This class is used to specify a DateRange, with start and end dates specified as 
strings formatted as YYYY-MM-DD.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{startDate}}{a character with format YYYY-MM-DD to be used as the date range's start date.}

\item{\code{endDate}}{a character with format YYYY-MM-DD to be used as the date range's end date.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DateRange$new()}}
\item \href{#method-toList}{\code{DateRange$toList()}}
\item \href{#method-asJSON}{\code{DateRange$asJSON()}}
\item \href{#method-clone}{\code{DateRange$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
creates a new instance, using the provided data param if provided.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DateRange$new(data = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{the data (as a json string or list) to initialize with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toList"></a>}}
\if{latex}{\out{\hypertarget{method-toList}{}}}
\subsection{Method \code{toList()}}{
returns the R6 class elements as a list for use in jsonlite::toJSON()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DateRange$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asJSON"></a>}}
\if{latex}{\out{\hypertarget{method-asJSON}{}}}
\subsection{Method \code{asJSON()}}{
returns the JSON string for this R6 class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DateRange$asJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DateRange$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
