% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortIncidence.R
\name{executeAnalysis}
\alias{executeAnalysis}
\title{Executes IR analysis given a design, options, and connection settings.}
\usage{
executeAnalysis(
  connectionDetails = NULL,
  connection = NULL,
  incidenceDesign,
  buildOptions,
  databaseName = "default"
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package. Either the \code{connection} or
\code{connectionDetails} argument should be specified.}

\item{connection}{A connection to the server containing the schema as created using the
\code{connect} function in the \code{DatabaseConnector} package.
Either the \code{connection} or \code{connectionDetails} argument
should be specified.}

\item{incidenceDesign}{A string object containing valid JSON that represents cohort incidence design}

\item{buildOptions}{the parameters to use in building the analysis queries, created by buildOptions()}
}
\value{
a \code{data.frame} containing the IR results
}
\description{
Executes IR analysis given a design, options, and connection settings.
}
