% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getCleanupSql}
\alias{getCleanupSql}
\title{Gets the sql that drops the results tables}
\usage{
getCleanupSql(useTempTables = F)
}
\arguments{
\item{useTempTables}{if true, then temp table notation will be used.}
}
\value{
SQL code in MS Sql Server dialect, if it's required to run analysis on another DBMS
        you have to use \code{\link[SqlRender]{translateSql}} function in the SqlRender package.
}
\description{
Gets the sql that drops the results tables
}
