% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getResultsDdl}
\alias{getResultsDdl}
\title{Gets the results schema DDL for Incidence Analysis}
\usage{
getResultsDdl()
}
\value{
SQL code in MS Sql Server dialect, if it's required to run analysis on another DBMS
        you have to use \code{\link[SqlRender]{translateSql}} function in the SqlRender package.
}
\description{
Gets the results schema DDL for Incidence Analysis
}
