% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{TimeAtRisk}
\alias{TimeAtRisk}
\title{R6 Class Representing an Time-at-Risk (TAR) definition}
\description{
The TimeAtRisk class, encapsulating the id, startWith, startOffset, endWith and endOffset
}
\details{
This class is used to specify a time-at-risk (TAR) definition. A TAR is defined by choosing
the start/end date of a cohort to start with (plus an offset), and a start/end date of the cohort
 to end with (plus an offset).

Note, when serializing with a library such as jsonlite, first call toList() on the R6 class
before calling jsonlite::toJSON().
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{id}}{an integer uniquely identifying this time at risk}

    \item{\code{startWith}}{the cohort date to start the time-at-risk.  Can be either "start" or "end".}

    \item{\code{startOffset}}{The number of days added to the date specified in startWith.}

    \item{\code{endWith}}{the cohort date to start the time-at-risk.  Can be either "start" or "end".}

    \item{\code{endOffset}}{The number of days added to the date specified in startWith.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TimeAtRisk-initialize}{\code{TimeAtRisk$new()}}
    \item \href{#method-TimeAtRisk-toList}{\code{TimeAtRisk$toList()}}
    \item \href{#method-TimeAtRisk-asJSON}{\code{TimeAtRisk$asJSON()}}
    \item \href{#method-TimeAtRisk-clone}{\code{TimeAtRisk$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAtRisk-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAtRisk-initialize}{}}}
\subsection{\code{TimeAtRisk$new()}}{
  creates a new instance, using the provided data param if provided.
The JSON takes the form: {"id":1,"start":{"dateField":"start","offset":1},"end":{"dateField":"start","offset":30}}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TimeAtRisk$new(data = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{the data (as a json string or list) to initialize with}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAtRisk-toList"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAtRisk-toList}{}}}
\subsection{\code{TimeAtRisk$toList()}}{
  returns the R6 class elements as a list for use in jsonlite::toJSON()
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TimeAtRisk$toList()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAtRisk-asJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAtRisk-asJSON}{}}}
\subsection{\code{TimeAtRisk$asJSON()}}{
  returns the JSON string for this R6 class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TimeAtRisk$asJSON()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAtRisk-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAtRisk-clone}{}}}
\subsection{\code{TimeAtRisk$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TimeAtRisk$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
